requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
alembic>=1.13.0
openskill>=5.0.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
jinja2>=3.1.3
python-multipart>=0.0.6
pytest>=7.4.0
pytest-asyncio>=0.21.0
python-telegram-bot>=20.0
python-levenshtein>=0.21.0
